extract into utils, format comments output
