+++ source/views/reddit/utils/format-count.ts
@@ -1,0 +1,10 @@
+export const formatCommentCount = (n: number): string => {
+	switch (n) {
+		case 0:
+			return 'Comments'
+		case 1:
+			return '1 comment'
+		default:
+			return `${n} comments`
+	}
+}
